delete and create post

+++ admin/delete-category.php
@@ -4,9 +4,18 @@ require 'config/database.php';
 if(isset($_GET['id'])){
     $id = filter_var($_GET['id'], FILTER_SANITIZE_SPECIAL_CHARS);
 
-    $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
-    $result = mysqli_query($connection, $query);
-    $_SESSION['delete-category-success'] = "Deleted with successsully";
+    $update_query = "UPDATE posts SET category_id=15 WHERE category_id=$id";
+    $update_result = mysqli_query($connection, $update_query);
+
+    if(!mysqli_errno($connection)){
+
+        $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
+        $result = mysqli_query($connection, $query);
+        $_SESSION['delete-category-success'] = "Deleted with successsully";
+
+    }
+
+  
 }
    header('location:manage-categories.php');
    die();
